chore: Split federation sql modules
Decrease module size by splitting different CRUD commands into dedicated
submodules.
